Remove `Filepath` class and integrate it into `LsCommand` class
